refactor: IntegrationTestSupport와 RestAssuredTestSupport 통합
