+++ src/test/java/roomescape/support/IntegrationTestSupport.java
@@ -1,17 +1,28 @@
 package roomescape.support;
 
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.jdbc.JdbcTestUtils;
 
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql(scripts = "/reset_test_data.sql")
 public abstract class IntegrationTestSupport {
 
+    @LocalServerPort
+    private int port;
+
     @Autowired
     private JdbcTemplate jdbcTemplate;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+    }
 
     protected int countRow(String tableName) {
         return JdbcTestUtils.countRowsInTable(jdbcTemplate, tableName);
